1.1.0 release quality code (excluding Linux AIO) -fixed ICF open/save dialogs due to a change starting in Vista -added access spec text to be displayed in status bar of GUI -adjusted perf screen graph scaling limits -cleaned up other code

+++ IOmeter/src/PageDisplay.h
@@ -253,8 +253,8 @@ class CPageDisplay:public CPropertyPage {
 	}
 
 	void OnBResultType(int which_button);
-	unsigned int GetMaxRange(Results * results, int result_type);
-	unsigned int ReportMaxRange(double max_range_needed);
+	double GetMaxRange(Results * results, int result_type);
+	double ReportMaxRange(double max_range_needed);
 
 	// Returns the appropriate result value in a double and a CString
 	BOOL GetDisplayData(Results * results, int result_type, double *result_value, CString * result_text);
